requirements.txt
absl-py==2.0.0
accelerate==0.27.2
aiohttp==3.9.3
aiosignal==1.3.1
altair==5.3.0
altgraph==0.17.3
annotated-types==0.6.0
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.6.0
asttokens==2.2.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==22.2.0
Augmentor==0.2.12
backcall==0.2.0
beautifulsoup4==4.11.2
bleach==6.0.0
blinker==1.7.0
blis==0.7.11
Brotli==1.1.0
cachetools==5.3.2
camelot-py==0.11.0
catalogue==2.0.10
certifi==2022.12.7
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.1.0
click==8.1.6
cloudpathlib==0.16.0
colorama==0.4.6
comm==0.1.2
confection==0.1.4
contourpy==1.0.7
cryptography==41.0.2
cv==1.0.0
cycler==0.11.0
cymem==2.0.8
dateparser==1.2.0
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
distro==1.8.0
Django==4.2.1
easyocr==1.6.2
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl
et-xmlfile==1.1.0
executing==1.2.0
fastjsonschema==2.16.2
filelock==3.10.7
flatbuffers==23.5.26
fonttools==4.39.4
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.2.0
gast==0.5.4
ghostscript==0.7
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.19.0
google-auth==2.23.3
google-auth-oauthlib==1.0.0
google-cloud-vision==3.7.2
google-pasta==0.2.0
googleapis-common-protos==1.63.1
googletrans==4.0.0rc1
grad-cam==1.5.0
graphviz==0.20.3
groundhog==0.11.2
grpcio==1.64.1
grpcio-status==1.62.2
h11==0.14.0
h2==3.2.0
h5py==3.10.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==1.0.4
httpx==0.27.0
huggingface-hub==0.21.3
hyperframe==5.2.0
idna==2.10
imageio==2.27.0
importlib-resources==6.0.0
inflate64==1.0.0
ipykernel==6.21.1
ipython==8.9.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.3.1
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-events==0.6.3
jupyter_client==8.0.2
jupyter_core==5.2.0
jupyter_server==2.2.1
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
keras==2.14.0
kiwisolver==1.4.4
langcodes==3.3.0
lazy_loader==0.2
libclang==16.0.6
lxml==5.2.2
Markdown==3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==2.0.5
ml-dtypes==0.2.0
mpmath==1.3.0
multidict==6.0.5
multivolumefile==0.2.3
murmurhash==1.0.10
mysql-connector-python==8.0.33
nbclassic==0.5.1
nbclient==0.7.2
nbconvert==7.2.9
nbformat==5.7.3
nest-asyncio==1.5.6
networkx==3.0
ninja==1.11.1
nltk==3.8.1
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.24.2
oauthlib==3.2.2
openai==0.28.0
opencv-python==4.7.0.72
opencv-python-headless==4.5.4.60
openpyxl==3.1.2
opt-einsum==3.3.0
packaging==23.0
pandas==2.0.1
pandocfilters==1.5.0
parso==0.8.3
pdf2image==1.16.3
pdfminer.six==20221105
pdfplumber==0.10.2
pefile==2023.2.7
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.0.0
plotly==5.18.0
preshed==3.0.9
prometheus-client==0.16.0
prompt-toolkit==3.0.36
proto-plus==1.23.0
protobuf==4.24.4
psutil==5.9.4
pure-eval==0.2.2
py7zr==0.21.0
pyarrow==15.0.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pybcj==1.0.2
pyclipper==1.3.0.post4
pycparser==2.21
pycryptodomex==3.20.0
pydantic==2.5.3
pydantic_core==2.14.6
pydeck==0.8.1b0
pygame==2.5.2
Pygments==2.14.0
pyinstaller==5.13.0
pyinstaller-hooks-contrib==2023.6
PyMuPDF==1.24.5
PyMuPDFb==1.24.3
pyngrok==7.1.6
pyparsing==3.0.9
pypdf==3.15.1
PyPDF2==2.11.1
pypdfium2==4.18.0
pyppmd==1.1.0
pyrsistent==0.19.3
pytesseract==0.3.10
python-bidi==0.4.2
python-dateutil==2.8.2
python-docx==1.1.2
python-json-logger==2.0.4
pytorch-gradcam==0.2.1
pytz==2023.3
PyWavelets==1.4.1
pywin32==305
pywin32-ctypes==0.2.2
pywinpty==2.0.10
PyYAML==6.0
pyzmq==25.0.0
pyzstd==0.15.9
qtconsole==5.4.0
QtPy==2.3.0
regex==2023.12.25
requests==2.28.2
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3986-validator==0.1.1
rich==13.7.1
rsa==4.9
safetensors==0.4.2
scikit-image==0.20.0
scikit-learn==1.3.0
scipy==1.11.1
seaborn==0.12.2
Send2Trash==1.8.0
shapely==2.0.1
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
sniffio==1.3.0
soundfile==0.12.1
soupsieve==2.3.2.post1
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.4.4
srsly==2.4.8
stack-data==0.6.2
streamlit==1.33.0
sweetviz==2.1.4
sympy==1.11.1
tabula==1.0.5
tabula-py==2.7.0
tabulate==0.9.0
tenacity==8.2.3
tensorboard==2.14.1
tensorboard-data-server==0.7.2
tensorflow==2.14.0
tensorflow-estimator==2.14.0
tensorflow-intel==2.14.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
terminado==0.17.1
texttable==1.7.0
thinc==8.2.2
threadpoolctl==3.2.0
tifffile==2023.3.21
tinycss2==1.2.1
tk==0.1.0
tokenizers==0.15.2
toml==0.10.2
toolz==0.12.1
torch==2.0.0
torchaudio==2.0.1
torchsummary==1.5.1
torchvision==0.15.1
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.38.2
ttach==0.0.3
typer==0.9.0
typing_extensions==4.9.0
tzdata==2023.3
tzlocal==5.2
uri-template==1.2.0
urllib3==1.26.15
wasabi==1.1.2
watchdog==4.0.0
wcwidth==0.2.6
weasel==0.3.4
webcolors==1.12
webencodings==0.5.1
websocket-client==1.5.1
Werkzeug==3.0.1
widgetsnbextension==4.0.5
wrapt==1.14.1
yarl==1.9.4
